Create enableApplyFiltersButton util
