+++ tests/e2e/specs/shopper/filter-products-by-attribute.test.ts
@@ -8,12 +8,8 @@ import {
 	insertBlock,
 	switchUserToAdmin,
 	publishPost,
-	ensureSidebarOpened,
 } from '@wordpress/e2e-test-utils';
-import {
-	selectBlockByName,
-	switchBlockInspectorTabWhenGutenbergIsInstalled,
-} from '@woocommerce/blocks-test-utils';
+import { selectBlockByName } from '@woocommerce/blocks-test-utils';
 
 /**
  * Internal dependencies
@@ -25,6 +21,7 @@ import {
 	useTheme,
 	waitForAllProductsBlockLoaded,
 	insertAllProductsBlock,
+	enableApplyFiltersButton,
 } from '../../utils';
 import { saveOrPublish } from '../../../utils';
 
@@ -36,8 +33,6 @@ const block = {
 		editor: {
 			firstAttributeInTheList:
 				'.woocommerce-search-list__list > li > label > input.woocommerce-search-list__item-input',
-			filterButtonToggle:
-				'//label[text()="Show \'Apply filters\' button"]/preceding-sibling::span[1]//input',
 			doneButton: '.wc-block-attribute-filter__selection > button',
 		},
 		frontend: {
@@ -48,8 +43,6 @@ const block = {
 			classicProductsList: '.products.columns-3 > li',
 			filter: "input[id='128gb']",
 			submitButton: '.wc-block-components-filter-submit-button',
-			XPathSubmitButton:
-				"//*[contains(@class,'wc-block-components-filter-submit-button')]",
 		},
 	},
 	urlSearchParamWhenFilterIsApplied:
@@ -192,21 +185,9 @@ describe( `${ block.name } Block`, () => {
 				postId: productCatalogTemplateId,
 			} );
 
-			await ensureSidebarOpened();
 			await selectBlockByName( block.slug );
-			await switchBlockInspectorTabWhenGutenbergIsInstalled( 'Settings' );
 
-			await page.evaluate( () => {
-				const toggle = document.querySelector(
-					'.components-toggle-control:last-child .components-form-toggle__input'
-				);
-				if ( ! toggle ) {
-					throw new Error( "'Apply filters' toggle not found" );
-				}
-				toggle.click();
-			} );
-
-			await page.waitForXPath( selectors.frontend.XPathSubmitButton );
+			await enableApplyFiltersButton();
 
 			await saveTemplate();
 			await goToShopPage();
@@ -319,16 +300,9 @@ describe( `${ block.name } Block`, () => {
 
 		it( 'should refresh the page only if the user clicks on button', async () => {
 			await page.goto( editorPageUrl );
-			await ensureSidebarOpened();
 			await selectBlockByName( block.slug );
-			await switchBlockInspectorTabWhenGutenbergIsInstalled( 'Settings' );
 
-			const [ filterButtonToggle ] = await page.$x(
-				block.selectors.editor.filterButtonToggle
-			);
-			await filterButtonToggle.click();
-
-			await canvas().waitForXPath( selectors.frontend.XPathSubmitButton );
+			await enableApplyFiltersButton();
 
 			await saveOrPublish();
 			await page.goto( frontedPageUrl );
